Add キノコを採ったらPlayerのサイズが変わるギミック実装　Playerの入力拒否関数実装

+++ project_Alice_in_Desktop/Assets/Scripts/Connector/Player/IPlayerStatusSentable.cs
@@ -4,7 +4,7 @@ namespace Connector
     {
         interface IPlayerStatusSentable 
         {
-            int ScaleMagnification { get; set; }
+            float ScaleMagnification { get; set; }
 
             void PlayerIsInput(bool flg);
         }
